add with option to call a includes scope before the sort scope
